feat(compteBO): acces a mon compte pour userBO 541
