+++ packages/backend/src/controllers/bo-authentication/email/login.js
@@ -55,13 +55,6 @@ module.exports = async function login(req, res, next) {
     );
   }
 
-  user.serviceCompetent =
-    user.territoireCode === "FRA"
-      ? "NAT"
-      : /^[0-9]+$/.test(user.territoireCode)
-        ? "DEP"
-        : "REG";
-
   try {
     const accessToken = jwt.sign(buildAccessToken(user), config.tokenSecret, {
       algorithm: "ES512",
