+++ packages/backend/src/controllers/bo-user/index.js
@@ -3,5 +3,6 @@ module.exports.getOne = require("./get-one");
 module.exports.getMe = require("./get-me");
 module.exports.create = require("./create");
 module.exports.update = require("./update");
+module.exports.updateMe = require("./update-me");
 module.exports.serviceCompetence = require("./service-competence");
 module.exports.verifyCompetence = require("./service-competence");
