+++ packages/backend/src/routes/bo-user.js
@@ -16,6 +16,8 @@ router.get("/", BOcheckJWT, BOcheckRoleCompte, BOUserController.list);
 router.get("/me", BOcheckJWT, BOUserController.getMe);
 // Renvoie les informations liées à l'utilisateur
 router.get("/:userId", BOcheckJWT, BOcheckRoleCompte, BOUserController.getOne);
+// Mise à jour de mes informations
+router.post("/me", BOcheckJWT, BOUserController.updateMe);
 // Création d'un utilisateur
 router.post(
   "/",
