+++ packages/frontend-bo/src/stores/user.js
@@ -26,6 +26,12 @@ export const useUserStore = defineStore("user", {
         if (user) {
           log.i("refreshProfile - DONE");
           this.user = user;
+          this.user.serviceCompetent =
+            this.user.territoireCode === "FRA"
+              ? "NAT"
+              : /^"\d"+$/.test(user.territoireCode)
+                ? "DEP"
+                : "REG";
         }
       } catch (err) {
         log.w("refreshProfile - DONE with error", err);
@@ -101,7 +107,6 @@ export const useUserStore = defineStore("user", {
     async getUser(id) {
       log.i("getUser - IN", { id });
       try {
-        // Appel du back pour la liste des utilisateurs
         const user = await $fetchBackend("/bo-user/" + id, {
           credentials: "include",
           method: "GET",
